+++ src/components/MyLayout.tsx
@@ -1,0 +1,4 @@
+import { Layout } from "react-admin";
+import { MainMenu } from "./MainMenu";
+
+export const MyLayout = (props) => <Layout {...props} menu={MainMenu} />;