+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * [feat] Add react proyect with react admin.
 * [docs] Update readme file.
+* [feat] Add Heroes III HOTA (system)